pyproject.toml
[project]
name = "neme-extractor"
version = "0.1.0"
description = "Extract anime character training crops from video using reference images, for character-LoRA training"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "typer>=0.12",
    "numpy>=1.26,<2.2",
    "Pillow>=10.4",
    "opencv-python>=4.10",
    "scenedetect>=0.6.5",
    "decord>=0.6.0",
    "tqdm>=4.66",
    "rich>=13.7",
    "pandas>=2.2",
    "pyarrow>=18.0",
]

[project.scripts]
neme-extractor = "neme_extractor.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neme_extractor"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.7.0",
]
gpu = [
    "torch>=2.4",
    "torchvision>=0.19",
    "dghs-imgutils>=0.19",
    "ultralytics>=8.3",
    "supervision>=0.24",
    "huggingface-hub>=0.26",
    "transformers>=4.45",
    "onnxruntime-gpu>=1.19",
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
markers = [
    "gpu: tests that require a CUDA GPU and may download model weights",
]
